Abstract a class responsible for adjusting the data cache size according to memory water level control.
