+++ be/src/agent/task_worker_pool.cpp
@@ -48,8 +48,8 @@
 #include "agent/report_task.h"
 #include "agent/resource_group_usage_recorder.h"
 #include "agent/task_signatures_manager.h"
+#include "cache/datacache.h"
 #include "cache/datacache_utils.h"
-#include "cache/local_cache.h"
 #include "common/status.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/workgroup/work_group.h"
@@ -839,7 +839,7 @@ void* ReportDataCacheMetricsTaskWorkerPool::_worker_thread_callback(void* arg_th
         request.__set_report_version(g_report_version.load(std::memory_order_relaxed));
 
         TDataCacheMetrics t_metrics{};
-        const LocalCache* cache = CacheEnv::GetInstance()->local_cache();
+        const LocalCache* cache = DataCache::GetInstance()->local_cache();
         if (cache != nullptr && cache->is_initialized()) {
             const DataCacheMetrics& metrics = cache->cache_metrics(0);
             DataCacheUtils::set_metrics_from_thrift(t_metrics, metrics);
