+++ be/src/cache/mem_space_monitor.cpp
@@ -1,0 +1,138 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "cache/mem_space_monitor.h"
+
+#include "common/config.h"
+#include "runtime/exec_env.h"
+#include "runtime/mem_tracker.h"
+#include "storage/page_cache.h"
+#include "util/gc_helper.h"
+#include "util/thread.h"
+
+#define SLEEP_IN_BG_WORKER(seconds)                                         \
+    int64_t left_seconds = (seconds);                                       \
+    while (!_stopped.load(std::memory_order_consume) && left_seconds > 0) { \
+        sleep(1);                                                           \
+        --left_seconds;                                                     \
+    }                                                                       \
+    if (_stopped.load(std::memory_order_consume)) {                         \
+        break;                                                              \
+    }
+
+namespace starrocks {
+
+void MemSpaceMonitor::start() {
+    _adjust_datacache_thread = std::thread([this] { _adjust_datacache_callback(); });
+    Thread::set_thread_name(_adjust_datacache_thread, "adjust_datacache");
+}
+
+void MemSpaceMonitor::stop() {
+    _stopped.store(true, std::memory_order_release);
+    if (_adjust_datacache_thread.joinable()) {
+        _adjust_datacache_thread.join();
+    }
+}
+
+void evict_pagecache(StoragePageCache* cache, int64_t bytes_to_dec, std::atomic<bool>& stoped) {
+    if (bytes_to_dec > 0) {
+        int64_t bytes = bytes_to_dec;
+        while (bytes >= GCBYTES_ONE_STEP) {
+            // Evicting 1GB of data takes about 1 second, check if process have been canceled.
+            if (UNLIKELY(stoped)) {
+                return;
+            }
+            cache->adjust_capacity(-GCBYTES_ONE_STEP, kcacheMinSize);
+            bytes -= GCBYTES_ONE_STEP;
+        }
+        if (bytes > 0) {
+            cache->adjust_capacity(-bytes, kcacheMinSize);
+        }
+    }
+}
+
+void MemSpaceMonitor::_adjust_datacache_callback() {
+    int64_t cur_period = config::pagecache_adjust_period;
+    int64_t cur_interval = config::auto_adjust_pagecache_interval_seconds;
+    std::unique_ptr<GCHelper> dec_advisor = std::make_unique<GCHelper>(cur_period, cur_interval, MonoTime::Now());
+    std::unique_ptr<GCHelper> inc_advisor = std::make_unique<GCHelper>(cur_period, cur_interval, MonoTime::Now());
+    auto* cache = _datacache->page_cache();
+    while (!_stopped.load(std::memory_order_consume)) {
+        SLEEP_IN_BG_WORKER(cur_interval);
+        if (!config::enable_auto_adjust_pagecache) {
+            continue;
+        }
+        if (config::disable_storage_page_cache) {
+            continue;
+        }
+        MemTracker* memtracker = GlobalEnv::GetInstance()->process_mem_tracker();
+        if (memtracker == nullptr || !memtracker->has_limit() || cache == nullptr) {
+            continue;
+        }
+        if (UNLIKELY(cur_period != config::pagecache_adjust_period ||
+                     cur_interval != config::auto_adjust_pagecache_interval_seconds)) {
+            cur_period = config::pagecache_adjust_period;
+            cur_interval = config::auto_adjust_pagecache_interval_seconds;
+            dec_advisor = std::make_unique<GCHelper>(cur_period, cur_interval, MonoTime::Now());
+            inc_advisor = std::make_unique<GCHelper>(cur_period, cur_interval, MonoTime::Now());
+            // We re-initialized advisor, just continue.
+            continue;
+        }
+
+        // Check config valid
+        int64_t memory_urgent_level = config::memory_urgent_level;
+        int64_t memory_high_level = config::memory_high_level;
+        if (UNLIKELY(!(memory_urgent_level > memory_high_level && memory_high_level >= 1 &&
+                       memory_urgent_level <= 100))) {
+            LOG(ERROR) << "memory water level config is illegal: memory_urgent_level=" << memory_urgent_level
+                       << " memory_high_level=" << memory_high_level;
+            continue;
+        }
+
+        int64_t memory_urgent = memtracker->limit() * memory_urgent_level / 100;
+        int64_t delta_urgent = memtracker->consumption() - memory_urgent;
+        int64_t memory_high = memtracker->limit() * memory_high_level / 100;
+        if (delta_urgent > 0) {
+            // Memory usage exceeds memory_urgent_level, reduce size immediately.
+            cache->adjust_capacity(-delta_urgent, kcacheMinSize);
+            size_t bytes_to_dec = dec_advisor->bytes_should_gc(MonoTime::Now(), memory_urgent - memory_high);
+            evict_pagecache(cache, static_cast<int64_t>(bytes_to_dec), _stopped);
+            continue;
+        }
+
+        int64_t delta_high = memtracker->consumption() - memory_high;
+        if (delta_high > 0) {
+            size_t bytes_to_dec = dec_advisor->bytes_should_gc(MonoTime::Now(), delta_high);
+            evict_pagecache(cache, static_cast<int64_t>(bytes_to_dec), _stopped);
+        } else {
+            auto ret = _datacache->get_storage_page_cache_limit();
+            if (!ret.ok()) {
+                LOG(ERROR) << "Failed to get storage page size: " << ret.status();
+                continue;
+            }
+            int64_t max_cache_size = std::max(ret.value(), kcacheMinSize);
+            int64_t cur_cache_size = cache->get_capacity();
+            if (cur_cache_size >= max_cache_size) {
+                continue;
+            }
+            int64_t delta_cache = std::min(max_cache_size - cur_cache_size, std::abs(delta_high));
+            size_t bytes_to_inc = inc_advisor->bytes_should_gc(MonoTime::Now(), delta_cache);
+            if (bytes_to_inc > 0) {
+                cache->adjust_capacity(bytes_to_inc);
+            }
+        }
+    }
+}
+
+} // namespace starrocks
