+++ be/src/http/action/update_config_action.cpp
@@ -43,7 +43,7 @@
 
 #include "agent/agent_common.h"
 #include "agent/agent_server.h"
-#include "cache/block_cache/block_cache.h"
+#include "cache/datacache.h"
 #include "common/configbase.h"
 #include "common/status.h"
 #include "exec/workgroup/scan_executor.h"
@@ -90,8 +90,8 @@ Status UpdateConfigAction::update_config(const std::string& name, const std::str
             return Status::OK();
         });
         _config_callback.emplace("storage_page_cache_limit", [&]() -> Status {
-            ASSIGN_OR_RETURN(int64_t cache_limit, CacheEnv::GetInstance()->get_storage_page_cache_limit());
-            cache_limit = CacheEnv::GetInstance()->check_storage_page_cache_limit(cache_limit);
+            ASSIGN_OR_RETURN(int64_t cache_limit, DataCache::GetInstance()->get_storage_page_cache_limit());
+            cache_limit = DataCache::GetInstance()->check_storage_page_cache_limit(cache_limit);
             StoragePageCache::instance()->set_capacity(cache_limit);
             return Status::OK();
         });
@@ -99,14 +99,14 @@ Status UpdateConfigAction::update_config(const std::string& name, const std::str
             if (config::disable_storage_page_cache) {
                 StoragePageCache::instance()->set_capacity(0);
             } else {
-                ASSIGN_OR_RETURN(int64_t cache_limit, CacheEnv::GetInstance()->get_storage_page_cache_limit());
-                cache_limit = CacheEnv::GetInstance()->check_storage_page_cache_limit(cache_limit);
+                ASSIGN_OR_RETURN(int64_t cache_limit, DataCache::GetInstance()->get_storage_page_cache_limit());
+                cache_limit = DataCache::GetInstance()->check_storage_page_cache_limit(cache_limit);
                 StoragePageCache::instance()->set_capacity(cache_limit);
             }
             return Status::OK();
         });
         _config_callback.emplace("datacache_mem_size", [&]() -> Status {
-            LocalCache* cache = CacheEnv::GetInstance()->local_cache();
+            LocalCache* cache = DataCache::GetInstance()->local_cache();
             if (cache == nullptr || !cache->is_initialized()) {
                 return Status::InternalError("Local cache is not initialized");
             }
@@ -121,7 +121,7 @@ Status UpdateConfigAction::update_config(const std::string& name, const std::str
             return cache->update_mem_quota(mem_size, true);
         });
         _config_callback.emplace("datacache_disk_size", [&]() -> Status {
-            LocalCache* cache = CacheEnv::GetInstance()->local_cache();
+            LocalCache* cache = DataCache::GetInstance()->local_cache();
             if (cache == nullptr || !cache->is_initialized()) {
                 return Status::InternalError("Local cache is not initialized");
             }
