+++ be/src/runtime/exec_env.cpp
@@ -40,10 +40,6 @@
 
 #include "agent/agent_server.h"
 #include "agent/master_info.h"
-#include "cache/block_cache/block_cache.h"
-#include "cache/disk_space_monitor.h"
-#include "cache/object_cache/lrucache_module.h"
-#include "cache/peer_cache_wrapper.h"
 #include "common/config.h"
 #include "common/configbase.h"
 #include "common/logging.h"
@@ -92,7 +88,6 @@
 #include "storage/lake/starlet_location_provider.h"
 #include "storage/lake/tablet_manager.h"
 #include "storage/lake/update_manager.h"
-#include "storage/page_cache.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet_schema_map.h"
 #include "storage/update_manager.h"
@@ -108,11 +103,6 @@
 
 #ifdef STARROCKS_JIT_ENABLE
 #include "exprs/jit/jit_engine.h"
-#endif
-
-#ifdef WITH_STARCACHE
-#include "cache/object_cache/starcache_module.h"
-#include "cache/starcache_wrapper.h"
 #endif
 
 namespace starrocks {
@@ -281,26 +271,6 @@ void GlobalEnv::_reset_tracker() {
     }
 }
 
-StatusOr<int64_t> CacheEnv::get_storage_page_cache_limit() {
-    return ParseUtil::parse_mem_spec(config::storage_page_cache_limit.value(), _global_env->process_mem_limit());
-}
-
-int64_t CacheEnv::check_storage_page_cache_limit(int64_t storage_cache_limit) {
-    if (storage_cache_limit > _global_env->process_mem_limit()) {
-        LOG(WARNING) << "Config storage_page_cache_limit is greater process memory limit, config="
-                     << config::storage_page_cache_limit.value() << ", memory=" << _global_env->process_mem_limit();
-        storage_cache_limit = _global_env->process_mem_limit();
-    }
-    if (!config::disable_storage_page_cache) {
-        if (storage_cache_limit < kcacheMinSize) {
-            LOG(WARNING) << "Storage cache limit is too small, use default size.";
-            storage_cache_limit = kcacheMinSize;
-        }
-        LOG(INFO) << "Set storage page cache size " << storage_cache_limit;
-    }
-    return storage_cache_limit;
-}
-
 std::shared_ptr<MemTracker> GlobalEnv::regist_tracker(MemTrackerType type, int64_t bytes_limit, MemTracker* parent) {
     auto mem_tracker = std::make_shared<MemTracker>(type, bytes_limit, MemTracker::type_to_label(type), parent);
     _mem_tracker_map[type] = mem_tracker;
@@ -332,208 +302,6 @@ bool parse_resource_str(const string& str, string* value) {
         }
     } else {
         return false;
-    }
-}
-
-CacheEnv* CacheEnv::GetInstance() {
-    static CacheEnv s_cache_env;
-    return &s_cache_env;
-}
-
-Status CacheEnv::init(const std::vector<StorePath>& store_paths) {
-    _global_env = GlobalEnv::GetInstance();
-    _store_paths = store_paths;
-
-    RETURN_IF_ERROR(_init_datacache());
-    RETURN_IF_ERROR(_init_starcache_based_object_cache());
-    RETURN_IF_ERROR(_init_lru_base_object_cache());
-    RETURN_IF_ERROR(_init_page_cache());
-
-    return Status::OK();
-}
-
-void CacheEnv::destroy() {
-    if (_disk_space_monitor != nullptr) {
-        _disk_space_monitor->stop();
-        _disk_space_monitor.reset();
-        LOG(INFO) << "disk space monitor stop successfully";
-    }
-
-    _page_cache.reset();
-    LOG(INFO) << "pagecache shutdown successfully";
-
-    _lru_based_object_cache.reset();
-    LOG(INFO) << "lru based object cache shutdown successfully";
-
-    _starcache_based_object_cache.reset();
-    LOG(INFO) << "starcache based object cache shutdown successfully";
-
-    _block_cache.reset();
-    LOG(INFO) << "datacache shutdown successfully";
-}
-
-Status CacheEnv::_init_starcache_based_object_cache() {
-#ifdef WITH_STARCACHE
-    if (_local_cache != nullptr && _local_cache->is_initialized()) {
-        auto* starcache = reinterpret_cast<StarCacheWrapper*>(_local_cache.get());
-        _starcache_based_object_cache = std::make_shared<StarCacheModule>(starcache->starcache_instance());
-    }
-#endif
-    return Status::OK();
-}
-
-Status CacheEnv::_init_lru_base_object_cache() {
-    ObjectCacheOptions options;
-    ASSIGN_OR_RETURN(int64_t storage_cache_limit, get_storage_page_cache_limit());
-    storage_cache_limit = check_storage_page_cache_limit(storage_cache_limit);
-    options.capacity = storage_cache_limit;
-
-    _lru_based_object_cache = std::make_shared<LRUCacheModule>(options);
-    LOG(INFO) << "object cache init successfully";
-    return Status::OK();
-}
-
-Status CacheEnv::_init_page_cache() {
-    _page_cache = std::make_shared<StoragePageCache>(_lru_based_object_cache.get());
-    _page_cache->init_metrics();
-    LOG(INFO) << "storage page cache init successfully";
-    return Status::OK();
-}
-
-Status CacheEnv::_init_datacache() {
-    _block_cache = std::make_shared<BlockCache>();
-
-    // When configured old `block_cache` configurations, use the old items for compatibility.
-    if (config::block_cache_enable) {
-        config::datacache_enable = true;
-        config::datacache_mem_size = std::to_string(config::block_cache_mem_size);
-        config::datacache_disk_size = std::to_string(config::block_cache_disk_size);
-        LOG(WARNING) << "The configuration items prefixed with `block_cache_` will be deprecated soon"
-                     << ", you'd better use the configuration items prefixed `datacache` instead!";
-    }
-
-#if !defined(WITH_STARCACHE)
-    if (config::datacache_enable) {
-        LOG(WARNING) << "No valid engines supported, skip initializing datacache module";
-        config::datacache_enable = false;
-    }
-#endif
-
-    if (config::datacache_enable) {
-#if defined(WITH_STARCACHE)
-        ASSIGN_OR_RETURN(auto cache_options, _init_cache_options());
-
-        // init starcache & disk monitor
-        _local_cache = std::make_shared<StarCacheWrapper>();
-        _disk_space_monitor = std::make_shared<DiskSpaceMonitor>(_local_cache.get());
-        RETURN_IF_ERROR(_disk_space_monitor->init(&cache_options.disk_spaces));
-        RETURN_IF_ERROR(_local_cache->init(cache_options));
-        _disk_space_monitor->start();
-
-        // init remote cache
-        _remote_cache = std::make_shared<PeerCacheWrapper>();
-        RETURN_IF_ERROR(_remote_cache->init(cache_options));
-
-        // init block cache
-        RETURN_IF_ERROR(_block_cache->init(cache_options, _local_cache, _remote_cache));
-
-        LOG(INFO) << "datacache init successfully";
-#endif
-    } else {
-        LOG(INFO) << "starts by skipping the datacache initialization";
-    }
-    return Status::OK();
-}
-
-StatusOr<CacheOptions> CacheEnv::_init_cache_options() {
-    CacheOptions cache_options;
-    RETURN_IF_ERROR(DataCacheUtils::parse_conf_datacache_mem_size(
-            config::datacache_mem_size, _global_env->process_mem_limit(), &cache_options.mem_space_size));
-
-    for (auto& root_path : _store_paths) {
-        // Because we have unified the datacache between datalake and starlet, we also need to unify the
-        // cache path and quota.
-        // To reuse the old cache data in `starlet_cache` directory, we try to rename it to the new `datacache`
-        // directory if it exists. To avoid the risk of cross disk renaming of a large amount of cached data,
-        // we do not automatically rename it when the source and destination directories are on different disks.
-        // In this case, users should manually remount the directories and restart them.
-        std::string datacache_path = root_path.path + "/datacache";
-        std::string starlet_cache_path = root_path.path + "/starlet_cache/star_cache";
-#ifdef USE_STAROS
-        if (config::datacache_unified_instance_enable) {
-            RETURN_IF_ERROR(DataCacheUtils::change_disk_path(starlet_cache_path, datacache_path));
-        }
-#endif
-        // Create it if not exist
-        Status st = FileSystem::Default()->create_dir_if_missing(datacache_path);
-        if (!st.ok()) {
-            LOG(ERROR) << "Fail to create datacache directory: " << datacache_path << ", reason: " << st.message();
-            return Status::InternalError("Fail to create datacache directory");
-        }
-
-        ASSIGN_OR_RETURN(int64_t disk_size, DataCacheUtils::parse_conf_datacache_disk_size(
-                                                    datacache_path, config::datacache_disk_size, -1));
-#ifdef USE_STAROS
-        // If the `datacache_disk_size` is manually set a positive value, we will use the maximum cache quota between
-        // dataleke and starlet cache as the quota of the unified cache. Otherwise, the cache quota will remain zero
-        // and then automatically adjusted based on the current avalible disk space.
-        if (config::datacache_unified_instance_enable && (!config::datacache_auto_adjust_enable || disk_size > 0)) {
-            ASSIGN_OR_RETURN(
-                    int64_t starlet_cache_size,
-                    DataCacheUtils::parse_conf_datacache_disk_size(
-                            datacache_path, fmt::format("{}%", config::starlet_star_cache_disk_size_percent), -1));
-            disk_size = std::max(disk_size, starlet_cache_size);
-        }
-#endif
-        cache_options.disk_spaces.push_back({.path = datacache_path, .size = static_cast<size_t>(disk_size)});
-    }
-
-    if (cache_options.disk_spaces.empty()) {
-        config::datacache_auto_adjust_enable = false;
-    }
-
-    // Adjust the default engine based on build switches.
-    if (config::datacache_engine == "") {
-#if defined(WITH_STARCACHE)
-        config::datacache_engine = "starcache";
-#endif
-    }
-    cache_options.block_size = config::datacache_block_size;
-    cache_options.max_flying_memory_mb = config::datacache_max_flying_memory_mb;
-    cache_options.max_concurrent_inserts = config::datacache_max_concurrent_inserts;
-    cache_options.enable_checksum = config::datacache_checksum_enable;
-    cache_options.enable_direct_io = config::datacache_direct_io_enable;
-    cache_options.enable_tiered_cache = config::datacache_tiered_cache_enable;
-    cache_options.skip_read_factor = config::datacache_skip_read_factor;
-    cache_options.scheduler_threads_per_cpu = config::datacache_scheduler_threads_per_cpu;
-    cache_options.enable_datacache_persistence = config::datacache_persistence_enable;
-    cache_options.inline_item_count_limit = config::datacache_inline_item_count_limit;
-    cache_options.engine = config::datacache_engine;
-    cache_options.eviction_policy = config::datacache_eviction_policy;
-    return cache_options;
-}
-
-void CacheEnv::try_release_resource_before_core_dump() {
-    std::set<std::string> modules;
-    bool release_all = false;
-    if (config::try_release_resource_before_core_dump.value() == "*") {
-        release_all = true;
-    } else {
-        SplitStringAndParseToContainer(StringPiece(config::try_release_resource_before_core_dump), ",",
-                                       &parse_resource_str, &modules);
-    }
-
-    auto need_release = [&release_all, &modules](const std::string& name) {
-        return release_all || modules.contains(name);
-    };
-
-    if (_page_cache != nullptr && need_release("data_cache")) {
-        _page_cache->set_capacity(0);
-    }
-    if (_local_cache != nullptr && _local_cache->available() && need_release("data_cache")) {
-        // TODO: Currently, block cache don't support shutdown now,
-        //  so here will temporary use update_mem_quota instead to release memory.
-        (void)_local_cache->update_mem_quota(0, false);
     }
 }
