+++ be/src/service/service_be/http_service.cpp
@@ -34,6 +34,7 @@
 
 #include "http_service.h"
 
+#include "cache/datacache.h"
 #include "fs/fs_util.h"
 #include "gutil/stl_util.h"
 #include "http/action/checksum_action.h"
@@ -63,12 +64,11 @@
 #include "http/http_method.h"
 #include "http/web_page_handler.h"
 #include "runtime/exec_env.h"
-#include "runtime/load_path_mgr.h"
 #include "util/starrocks_metrics.h"
 
 namespace starrocks {
 
-HttpServiceBE::HttpServiceBE(CacheEnv* cache_env, ExecEnv* env, int port, int num_threads)
+HttpServiceBE::HttpServiceBE(DataCache* cache_env, ExecEnv* env, int port, int num_threads)
         : _cache_env(cache_env),
           _env(env),
           _ev_http_server(new EvHttpServer(port, num_threads)),
