+++ be/src/util/system_metrics.cpp
@@ -42,7 +42,7 @@
 #include <cstdio>
 #include <memory>
 
-#include "cache/local_cache.h"
+#include "cache/datacache.h"
 #ifdef USE_STAROS
 #include "fslib/star_cache_handler.h"
 #endif
@@ -301,7 +301,7 @@ void SystemMetrics::_update_datacache_mem_tracker() {
     int64_t datacache_mem_bytes = 0;
     auto* datacache_mem_tracker = GlobalEnv::GetInstance()->datacache_mem_tracker();
     if (datacache_mem_tracker) {
-        LocalCache* local_cache = CacheEnv::GetInstance()->local_cache();
+        LocalCache* local_cache = DataCache::GetInstance()->local_cache();
         if (local_cache != nullptr && local_cache->is_initialized()) {
             auto datacache_metrics = local_cache->cache_metrics(0);
             datacache_mem_bytes = datacache_metrics.mem_used_bytes + datacache_metrics.meta_used_bytes;
