+++ be/test/cache/object_cache/starcache_module_test.cpp
@@ -71,7 +71,7 @@ protected:
 
 void StarCacheModuleTest::_init_local_cache() {
     CacheOptions options = TestCacheUtils::create_simple_options(256 * KB, _mem_quota);
-    options.disk_spaces.push_back({.path = cache_dir, .size = 50 * MB});
+    options.dir_spaces.push_back({.path = cache_dir, .size = 50 * MB});
 
     _local_cache = std::make_shared<StarCacheWrapper>();
     ASSERT_OK(_local_cache->init(options));
